# backend/routes/api.yaml
swagger: "2.0"
info:
  title: BrainBeatsWeb API 
  description: This is a BrainBeatsWeb API documentation form
  version: 1.0
  licenses:
    name: Apache 2.0
    url: "http://www.apache.org/liscences/LIS"
  servers:
    -url: http://api.brainbeatz.xyz #https://api.brainbeats.dev
    description: Local server
schemes: [http] #[https]
...
paths:
...
    /api/scripts/createScript:
      post:
        tags:
          - Scripts
        summary: Create script at user profile
        operationId: createScript
        description: Create a script for a user by username
        consumes: Scripts
         - application/json
        produces:
        - application/json
        parameters:
        - in: body
          name: Create script
          required: true
          description: User Script
          schema:
            $ref: '#/definitions/createScript'
        responses:
          200:
            description: OK
          404:
            description: URL Not Found

  # Get user tracks by username
...
definitions:
...
    createScript:
      required:
      - userID
      - title
      - visibility
      properties:
        userID:
          type: string
          example: 227d4d04-927b-421e-8d95-6eb91f05af05
        title:
          type: string
          example: Track title
        visibility:
          type: boolean
          example: True



  # ***** User Playlists ***** #
...